// libs/westfield-rise-contentful-client/@types/generated/contentful.d.ts
export interface IAccordionFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title?: string | undefined;

  /** Accordion Items */
  accordionItems?: IAccordionItem[] | undefined;

  /** with Number in front */
  withNumber: boolean;

  /** Index of opened Accordion */
  indexOfOpened?: number | undefined;
}

/** Accepts: Accordion Item */

export interface IAccordion extends Entry<IAccordionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'accordion';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IAcquisition extends Entry<IAcquisitionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'acquisition';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IBodyCopyFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title?: string | undefined;

  /** Body Copy Text */
  bodyCopyText: Document;
}

/** Rich Text editor component */

export interface IBodyCopy extends Entry<IBodyCopyFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'bodyCopy';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IBodyCopyComboList extends Entry<IBodyCopyComboListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'bodyCopyComboList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface ICallToAction extends Entry<ICallToActionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'callToAction';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICardsFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title?: string | undefined;

  /** Card List */
  cardList: IListCard[];

  /** With Motion Background */
  withMotionBackground?: boolean | undefined;
}

/** List of: Case study cards / service cards / Insight cards */

export interface ICards extends Entry<ICardsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'cards';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface ICaseStudy extends Entry<ICaseStudyFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'caseStudy';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICaseStudyCarouselFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title?: string | undefined;

  /** Case Study List */
  caseStudyList: IImageStatistics[];
}

export interface ICaseStudyCarousel extends Entry<ICaseStudyCarouselFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'caseStudyCarousel';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface ICaseStudyListFields {
  /** Internal Name */
  internalName: string;

  /** CaseStudies */
  caseStudies?: ICaseStudyCardSmall[] | undefined;
}

export interface ICaseStudyList extends Entry<ICaseStudyListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'caseStudyList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IClientCarouselFields {
  /** Internal Name */
  internalName: string;

  /** Logo List */
  logoList: Asset[];
}

export interface IClientCarousel extends Entry<IClientCarouselFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'clientCarousel';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IDownloadPdfFields {
  /** Internal Name */
  internalName: string;

  /** Text */
  text: string;

  /** Call to Action */
  callToAction: ICallToAction;
}

export interface IDownloadPdf extends Entry<IDownloadPdfFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'downloadPdf';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IFilterableSortableListFields {
  /** Internal Name */
  internalName: string;

  /** List Item Type */
  listItemType?: ICaseStudy | IPageKeyInformation | undefined;
}

export interface IFilterableSortableList
  extends Entry<IFilterableSortableListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'filterableSortableList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IFooterMenu extends Entry<IFooterMenuFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'footerMenu';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IHeaderLargeFields {
  /** Internal Name */
  internalName: string;

  /** Type */
  type: 'text' | 'image' | 'video';

  /** Content */
  content?: IHeaderText | IVimeoComponent | undefined;

  /** Image */
  image?: Asset | undefined;
}

/** Accepts: Header Text, Vimeo Component or Image */

export interface IHeaderLarge extends Entry<IHeaderLargeFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'headerLarge';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IHeaderSmallFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title: string;

  /** Image */
  image?: Asset | undefined;
}

export interface IHeaderSmall extends Entry<IHeaderSmallFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'headerSmall';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IHomePageHero extends Entry<IHomePageHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'homePageHero';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IImageCarouselFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Type */
  type?: 'default' | 'people' | 'news-insights' | undefined;

  /** Title */
  title?: string | undefined;

  /** Call to Action */
  callToAction?: ICallToAction | undefined;

  /** Image List */
  imageList?: IImageCarouselItem[] | undefined;
}

export interface IImageCarousel extends Entry<IImageCarouselFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'imageCarousel';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IImageText extends Entry<IImageTextFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'imageText';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IImageTextListFields {
  /** Internal Name */
  internalName?: string | undefined;

  /** Title */
  title?: string | undefined;

  /** Image Text List */
  imageTextList?: IImageText[] | undefined;
}

/** Accepts: Image Text components */

export interface IImageTextList extends Entry<IImageTextListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'imageTextList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IKeyFactListFields {
  /** Internal Name */
  internalName: string;

  /** Key Fact List */
  keyFactList: IKeyFacts[];
}

export interface IKeyFactList extends Entry<IKeyFactListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'keyFactList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IMap extends Entry<IMapFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'map';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface INavigationMenu extends Entry<INavigationMenuFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'navigationMenu';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IPageFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title: string;

  /** slug */
  slug: string;

  /** Body */
  body: (
    | IAccordion
    | IHeaderLarge
    | IQuote
    | IVimeoComponent
    | IStatisticsList
    | ICaseStudyList
    | IBodyCopy
    | IImageCarousel
    | IImageTextList
    | ICaseStudy
    | ICards
    | IPageKeyInformation
    | IFooterMenu
    | IImageText
    | ICallToAction
    | IClientCarousel
    | IDownloadPdf
    | IHeaderSmall
    | IMap
    | IHomePageHero
    | IKeyFactList
    | IAcquisition
    | IReferenceDescription
    | IBodyCopyComboList
    | ICaseStudyCarousel
    | INavigationMenu
    | IFilterableSortableList
  )[];

  /** Related Pages */
  relatedPages?: IPage[] | undefined;

  /** SEO metadata */
  seoMetadata?: ISeoMetaTags | undefined;

  /** Theme */
  theme:
    | 'canary'
    | 'candy'
    | 'peacock'
    | 'scarlet'
    | 'tangerine'
    | 'ultraviolet';

  /** Has Gradient Background */
  hasGradient?: boolean | undefined;

  /** Navigation has White Background */
  whiteNav?: boolean | undefined;

  /** tecComponents */
  tecComponents?: Record<string, any> | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'page';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IPageKeyInformation extends Entry<IPageKeyInformationFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'pageKeyInformation';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IQuoteFields {
  /** Internal Name */
  internalName: string;

  /** Quote */
  quote?: string | undefined;

  /** Author */
  author: string;

  /** Job Title */
  jobTitle: string;

  /** Background Image */
  backgroundImage?: Asset | undefined;
}

export interface IQuote extends Entry<IQuoteFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'quote';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IReferenceDescription
  extends Entry<IReferenceDescriptionFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'referenceDescription';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface ISeoMetaTags extends Entry<ISeoMetaTagsFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'seoMetaTags';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}

export interface IStatisticsListFields {
  /** Internal Name */
  internalName: string;

  /** Title */
  title?: string | undefined;

  /** Statistics List */
  statisticsList: IKpiComponent[];

  /** withAnimation */
  withAnimation?: boolean | undefined;
}

/** A list of KPI Components */

export interface IStatisticsList extends Entry<IStatisticsListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'statisticsList';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}
// ...
export interface IVimeoComponentFields {
  /** Internal Name */
  internalName: string;

  /** VimeoId */
  vimeoId: string;

  /** Image */
  image?: Asset | undefined;

  /** Full Width */
  fullWidth?: boolean | undefined;
}

/** Usable only within Header Large */

export interface IVimeoComponent extends Entry<IVimeoComponentFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: 'vimeoComponent';
        linkType: 'ContentType';
        type: 'Link';
      };
    };
  };
}